Add re-propose feature

+++ server/routes/controllers/payment.js
@@ -11,16 +11,18 @@ sgMail.setApiKey(config.SENDGRID_API_KEY);
 
 const stripe = require('stripe')(config.STRIPE_SK)
 
+const REQUEST_DECLINED_BY_OWNER = 'request_declined_by_owner'
+const REQUEST_ACCEPTED_BY_OWNER = 'request_accepted_by_owner'
+const REQUEST_DECLINED_BY_USER = 'request_declined_by_user'
+const REQUEST_ACCEPTED_BY_USER = 'request_accepted_by_user'
 
-const REQUEST_DECLINED = 'request_declined'
-const REQUEST_ACCEPTED = 'request_accepted'
 
 function sendEmailTo(sendTo, sendMsg, booking, hostname) {
     let msg = {}
     const startAt = moment(booking.startAt).tz("Asia/Tokyo").format("YYYY/MM/DD/HH:mm")
     const endAt = moment(booking.endAt).tz("Asia/Tokyo").format("HH:mm")
 
-    if(sendMsg == REQUEST_DECLINED) {
+    if(sendMsg == REQUEST_DECLINED_BY_OWNER) {
         msg = {
             to: sendTo,
             from: "noreply@ap-trainer.com",
@@ -33,7 +35,7 @@ function sendEmailTo(sendTo, sendMsg, booking, hostname) {
                 + '\n\n\n\n'
                 + 'Anytime Personal Trainer.inc'
         }
-    } else if (sendMsg == REQUEST_ACCEPTED) {
+    } else if (sendMsg == REQUEST_ACCEPTED_BY_OWNER) {
         msg = {
             to: sendTo,
             from: "noreply@ap-trainer.com",
@@ -42,6 +44,32 @@ function sendEmailTo(sendTo, sendMsg, booking, hostname) {
                 + '日時：' + startAt + ' 〜 ' + endAt + ' \n\n'
                 + '場所：' + booking.rental.province + '\n\n'
                 + 'これ以降のキャンセルはできません。時間に余裕を持って目的地に到着されるようお願いいたします。'
+                + '\n\n\n\n'
+                + 'Anytime Personal Trainer.inc'
+        }
+    } else if (sendMsg == REQUEST_ACCEPTED_BY_USER) {
+        msg = {
+            to: sendTo,
+            from: "noreply@ap-trainer.com",
+            subject: "[予約確定]「" + booking.rental.rentalname + " Trainer」への予約日時変更リクエストが受理されました！",
+            text: 'おめでとうございます！「' + booking.rental.rentalname + ' Trainer」の予約日時変更リクエストが受理されました！\n\n' 
+                + '日時：' + startAt + ' 〜 ' + endAt + ' \n\n'
+                + '場所：' + booking.rental.province + '\n\n'
+                + 'これ以降のキャンセルはできません。生徒に満足される最高のおもてなしをしましょう！'
+                + '\n\n\n\n'
+                + 'Anytime Personal Trainer.inc'
+        }
+    } else if (sendMsg == REQUEST_DECLINED_BY_USER) {
+        msg = {
+            to: sendTo,
+            from: "noreply@ap-trainer.com",
+            subject: "「" + booking.rental.rentalname + " Trainer」への予約日時変更リクエストは受理されませんでした",
+            text: "商品名：" + booking.rental.rentalname + " \n\n"
+                + "日時：" + startAt + ' 〜 ' + endAt + " \n\n"
+                + "場所：" + booking.rental.province + "\n\n"
+                + "への予約日時変更リクエストは受理されませんでした。\n\n"
+                + "せっかくの予約機会を逃さぬよう、受付不可日はなるべく早めに予約ブロック機能で事前にブロックしておきましょう。\n\n"
+                + "予約ブロック機能はTrainer管理ページ内で設定できます。"
                 + '\n\n\n\n'
                 + 'Anytime Personal Trainer.inc'
         }
@@ -60,7 +88,10 @@ exports.getPendingPayments = function(req, res) {
             .populate({
                 // populate both 'booking' and 'rental'
                 path: 'booking',
-                populate: {path: 'rental'}
+                populate: {
+                    path: 'rental',
+                    populate: {path: 'bookings'} // This is using for repropose booking date from rental owner.
+                }
             })
             .populate('fromUser')
             .exec(function(err, foundPayments) {
@@ -92,10 +123,11 @@ exports.getPaidPayments = function(req, res) {
 }
 
 exports.acceptPayment = function(req, res) {
-    const payment = req.body
+    const payment  = req.body
     const user = res.locals.user
 
     Payment.findById(payment._id)
+            .populate('fromUser')
             .populate('toUser')
             .populate({
                 // populate both 'booking' and 'rental'
@@ -107,41 +139,51 @@ exports.acceptPayment = function(req, res) {
             return res.status(422).send({errors: normalizeErrors(err.errors)})
         }
 
-        if(foundPayment.status == 'pending' && user.id == foundPayment.toUser.id) {
-            const booking = foundPayment.booking
+        if(foundPayment.status !== 'pending') {
+            return res.status(422).send({errors: [{title: "Invalid request!", detail: "This payment is already paid or canseled!"}]})
+        }
 
-            // It will charge to booking user when owner accept payment.
-            // This charge is not directly going to rental owner, but paid to platform owner.
-            // Platform owner have to pay to rental owner manually considering commision rate.
-            const charge = await stripe.charges.create({
-                amount: booking.totalPrice,
-                currency: 'jpy',
-                customer: payment.fromStripeCustomerId
-            })
+        if(!(foundPayment.toUser.id === user.id || foundPayment.fromUser.id === user.id)) {
+            return res.status(422).send({errors: [{title: "Invalid user!", detail: "You are not rental owner or booking user!"}]})
+        }
 
-            if(charge) {
+        // It will charge to booking user when owner accept payment. 
+        // Or, booking user accept reproposal date from rental owner.
+        // This charge is not directly going to rental owner, but paid to platform owner.
+        // Platform owner have to pay to rental owner manually considering commision rate.
+        const booking = foundPayment.booking
+        const charge = await stripe.charges.create({
+            amount: booking.totalPrice,
+            currency: 'jpy',
+            customer: foundPayment.fromStripeCustomerId
+        })
+
+        if(charge) {
+            foundPayment.charge = charge
+            foundPayment.status = 'paid'
+            foundPayment.paidAt = new Date()
+            foundPayment.save(function(err) {
+                if(err) {
+                    return res.status(422).send({errors: normalizeErrors(err.errors)})
+                }
                 Booking.updateMany({ _id: booking}, { status: 'active'}, function(){})
 
-                foundPayment.charge = charge
-                foundPayment.status = 'paid'
-                foundPayment.paidAt = new Date()
-                foundPayment.save(function(err) {
-                    if(err) {
-                        return res.status(422).send({errors: normalizeErrors(err.errors)})
-                    }
-                    sendEmailTo(payment.fromUser.email, REQUEST_ACCEPTED, booking, req.hostname)
-                    return res.json({ status: 'paid'})
+                if(foundPayment.toUser.id === user.id) {
+                    sendEmailTo(foundPayment.fromUser.email, REQUEST_ACCEPTED_BY_OWNER, booking, req.hostname)
+                }
+                if (foundPayment.fromUser.id === user.id) {
+                    sendEmailTo(foundPayment.toUser.email, REQUEST_ACCEPTED_BY_USER, booking, req.hostname)
+                }
+                return res.json({ status: 'paid'})
 
-
-                    // User.updateMany({_id: foundPayment.toUser}, { $inc: {revenue: foundPayment.amount}}, function(err, user) {
-                    //     if(err) {
-                    //         return res.status(422).send({errors: normalizeErrors(err.errors)})
-                    //     }
-                    //     sendEmailTo(payment.fromUser.email, ACCEPTED)
-                    //     return res.json({ status: 'paid'})
-                    // })
-                })
-            }
+                // User.updateMany({_id: foundPayment.toUser}, { $inc: {revenue: foundPayment.amount}}, function(err, user) {
+                //     if(err) {
+                //         return res.status(422).send({errors: normalizeErrors(err.errors)})
+                //     }
+                //     sendEmailTo(payment.fromUser.email, ACCEPTED)
+                //     return res.json({ status: 'paid'})
+                // })
+            })
         }
     })
 }
@@ -150,17 +192,18 @@ exports.declinePayment = function(req, res) {
     const payment = req.body
     const { booking } = req.body
 
+    if(booking.status === 'active') {
+        return res.status(422).send({errors: [{title: "Invalid request!", detail: "Cannot delete active bookings payment!"}]})
+    }
+
     Booking.deleteOne({id: booking._id}, (err) => { // Delete Booking
         if(err) {
             return res.status(422).send({errors: normalizeErrors(err.errors)})
         }
-
-        sendEmailTo(payment.fromUser.email, REQUEST_DECLINED, booking, req.hostname)
+        sendEmailTo(payment.fromUser.email, REQUEST_DECLINED_BY_OWNER, booking, req.hostname)
 
         Payment.updateMany({_id: payment._id}, {status: 'declined'}, function(){})
         Rental.updateMany({_id: booking.rental}, {$pull: {bookings: booking._id}}, ()=>{}) // Delete Booking from Rental
-
         return res.json({status: 'deleted'})
     })
-
 }
